Apply Strategy pattern for syntax highlighting
