+++ FastColoredTextBox/Language/CustomLanguage.cs
@@ -1,0 +1,196 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Globalization;
+using System.IO;
+using System.Text;
+using System.Text.RegularExpressions;
+using System.Xml;
+using FastColoredTextBoxNS.Language2;
+
+namespace FastColoredTextBoxNS.Language2
+{
+	public class CustomLanguage : ILanguage, IDisposable
+	{
+        protected readonly Dictionary<string, SyntaxDescriptor> descByXMLfileNames =
+            new Dictionary<string, SyntaxDescriptor>();
+
+		public string XMLdescriptionFile { get; set; }
+		public void AutoIndentNeeded(FastColoredTextBox fctb, AutoIndentEventArgs args) {
+			fctb.CalcAutoIndentShiftByCodeFolding(fctb, args);
+		}
+
+		public void SyntaxHighlight(Range range) {
+            SyntaxDescriptor desc = null;
+            if (!descByXMLfileNames.TryGetValue(XMLdescriptionFile, out desc))
+            {
+                var doc = new XmlDocument();
+                string file = XMLdescriptionFile;
+                if (!File.Exists(file))
+                    file = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, Path.GetFileName(file));
+
+                doc.LoadXml(File.ReadAllText(file));
+                desc = ParseXmlDescription(doc);
+                descByXMLfileNames[XMLdescriptionFile] = desc;
+            }
+
+            //set style order
+            range.tb.ClearStylesBuffer();
+            for (int i = 0; i < desc.styles.Count; i++)
+                range.tb.Styles[i] = desc.styles[i];
+            //brackets
+            char[] oldBrackets = RememberBrackets(range.tb);
+            range.tb.LeftBracket = desc.leftBracket;
+            range.tb.RightBracket = desc.rightBracket;
+            range.tb.LeftBracket2 = desc.leftBracket2;
+            range.tb.RightBracket2 = desc.rightBracket2;
+            //clear styles of range
+            range.ClearStyle(desc.styles.ToArray());
+            //highlight syntax
+            foreach (RuleDesc rule in desc.rules)
+                range.SetStyle(rule.style, rule.Regex);
+            //clear folding
+            range.ClearFoldingMarkers();
+            //folding markers
+            foreach (FoldingDesc folding in desc.foldings)
+                range.SetFoldingMarkers(folding.startMarkerRegex, folding.finishMarkerRegex, folding.options);
+
+            //
+            RestoreBrackets(range.tb, oldBrackets);
+		}
+		        protected char[] RememberBrackets(FastColoredTextBox tb)
+        {
+            return new[] { tb.LeftBracket, tb.RightBracket, tb.LeftBracket2, tb.RightBracket2 };
+        }
+        protected void RestoreBrackets(FastColoredTextBox tb, char[] oldBrackets)
+        {
+            tb.LeftBracket = oldBrackets[0];
+            tb.RightBracket = oldBrackets[1];
+            tb.LeftBracket2 = oldBrackets[2];
+            tb.RightBracket2 = oldBrackets[3];
+        }
+        protected static FoldingDesc ParseFolding(XmlNode foldingNode)
+        {
+            var folding = new FoldingDesc();
+            //regex
+            folding.startMarkerRegex = foldingNode.Attributes["start"].Value;
+            folding.finishMarkerRegex = foldingNode.Attributes["finish"].Value;
+            //options
+            XmlAttribute optionsA = foldingNode.Attributes["options"];
+            if (optionsA != null)
+                folding.options = (RegexOptions)Enum.Parse(typeof(RegexOptions), optionsA.Value);
+
+            return folding;
+        }
+        public static SyntaxDescriptor ParseXmlDescription(XmlDocument doc)
+        {
+            var desc = new SyntaxDescriptor();
+            XmlNode brackets = doc.SelectSingleNode("doc/brackets");
+            if (brackets != null)
+            {
+                if (brackets.Attributes["left"] == null || brackets.Attributes["right"] == null ||
+                    brackets.Attributes["left"].Value == "" || brackets.Attributes["right"].Value == "")
+                {
+                    desc.leftBracket = '\x0';
+                    desc.rightBracket = '\x0';
+                }
+                else
+                {
+                    desc.leftBracket = brackets.Attributes["left"].Value[0];
+                    desc.rightBracket = brackets.Attributes["right"].Value[0];
+                }
+
+                if (brackets.Attributes["left2"] == null || brackets.Attributes["right2"] == null ||
+                    brackets.Attributes["left2"].Value == "" || brackets.Attributes["right2"].Value == "")
+                {
+                    desc.leftBracket2 = '\x0';
+                    desc.rightBracket2 = '\x0';
+                }
+                else
+                {
+                    desc.leftBracket2 = brackets.Attributes["left2"].Value[0];
+                    desc.rightBracket2 = brackets.Attributes["right2"].Value[0];
+                }
+
+                if (brackets.Attributes["strategy"] == null || brackets.Attributes["strategy"].Value == "")
+                    desc.bracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+                else
+                    desc.bracketsHighlightStrategy = (BracketsHighlightStrategy)Enum.Parse(typeof(BracketsHighlightStrategy), brackets.Attributes["strategy"].Value);
+            }
+
+            var styleByName = new Dictionary<string, Style>();
+
+            foreach (XmlNode style in doc.SelectNodes("doc/style"))
+            {
+                Style s = ParseStyle(style);
+                styleByName[style.Attributes["name"].Value] = s;
+                desc.styles.Add(s);
+            }
+            foreach (XmlNode rule in doc.SelectNodes("doc/rule"))
+                desc.rules.Add(ParseRule(rule, styleByName));
+            foreach (XmlNode folding in doc.SelectNodes("doc/folding"))
+                desc.foldings.Add(ParseFolding(folding));
+
+            return desc;
+        }
+        protected static RuleDesc ParseRule(XmlNode ruleNode, Dictionary<string, Style> styles)
+        {
+            var rule = new RuleDesc();
+            rule.pattern = ruleNode.InnerText;
+            //
+            XmlAttribute styleA = ruleNode.Attributes["style"];
+            XmlAttribute optionsA = ruleNode.Attributes["options"];
+            //Style
+            if (styleA == null)
+                throw new Exception("Rule must contain style name.");
+            if (!styles.ContainsKey(styleA.Value))
+                throw new Exception("Style '" + styleA.Value + "' is not found.");
+            rule.style = styles[styleA.Value];
+            //options
+            if (optionsA != null)
+                rule.options = (RegexOptions)Enum.Parse(typeof(RegexOptions), optionsA.Value);
+
+            return rule;
+        }
+        protected static Style ParseStyle(XmlNode styleNode)
+        {
+            XmlAttribute typeA = styleNode.Attributes["type"];
+            XmlAttribute colorA = styleNode.Attributes["color"];
+            XmlAttribute backColorA = styleNode.Attributes["backColor"];
+            XmlAttribute fontStyleA = styleNode.Attributes["fontStyle"];
+            XmlAttribute nameA = styleNode.Attributes["name"];
+            //colors
+            SolidBrush foreBrush = null;
+            if (colorA != null)
+                foreBrush = new SolidBrush(ParseColor(colorA.Value));
+            SolidBrush backBrush = null;
+            if (backColorA != null)
+                backBrush = new SolidBrush(ParseColor(backColorA.Value));
+            //fontStyle
+            FontStyle fontStyle = FontStyle.Regular;
+            if (fontStyleA != null)
+                fontStyle = (FontStyle)Enum.Parse(typeof(FontStyle), fontStyleA.Value);
+
+            return new TextStyle(foreBrush, backBrush, fontStyle);
+        }
+
+        protected static Color ParseColor(string s)
+        {
+            if (s.StartsWith("#"))
+            {
+                if (s.Length <= 7)
+                    return Color.FromArgb(255,
+                                          Color.FromArgb(Int32.Parse(s.Substring(1), NumberStyles.AllowHexSpecifier)));
+                else
+                    return Color.FromArgb(Int32.Parse(s.Substring(1), NumberStyles.AllowHexSpecifier));
+            }
+            else
+                return Color.FromName(s);
+        }
+        public void Dispose()
+        {
+            foreach (SyntaxDescriptor desc in descByXMLfileNames.Values)
+                desc.Dispose();
+        }
+	}
+}
