+++ FastColoredTextBox/Language/Lua.cs
@@ -1,0 +1,108 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Text.RegularExpressions;
+using FastColoredTextBoxNS.Language2;
+
+namespace FastColoredTextBoxNS.Language2
+{
+	public class Lua : ILanguage
+	{
+		private readonly Style StringStyle = LanguageStyles.Brown;
+		private readonly Style CommentStyle = LanguageStyles.Green;
+		private readonly Style NumberStyle = LanguageStyles.Magenta;
+		private readonly Style KeywordStyle = LanguageStyles.BlueBold;
+		private readonly Style FunctionsStyle = LanguageStyles.Maroon;
+
+        protected Regex LuaCommentRegex1,
+                      LuaCommentRegex2,
+                      LuaCommentRegex3;
+
+        protected Regex LuaKeywordRegex;
+        protected Regex LuaNumberRegex;
+        protected Regex LuaStringRegex;
+        protected Regex LuaFunctionsRegex;
+
+
+		public Lua() {
+			InitLuaRegex();
+		}
+        protected void InitLuaRegex()
+        {
+            LuaStringRegex = new Regex(@"""""|''|"".*?[^\\]""|'.*?[^\\]'", Parsing.RegexCompiledOption);
+            LuaCommentRegex1 = new Regex(@"--.*$", RegexOptions.Multiline | Parsing.RegexCompiledOption);
+            LuaCommentRegex2 = new Regex(@"(--\[\[.*?\]\])|(--\[\[.*)", RegexOptions.Singleline | Parsing.RegexCompiledOption);
+            LuaCommentRegex3 = new Regex(@"(--\[\[.*?\]\])|(.*\]\])",
+                                             RegexOptions.Singleline | RegexOptions.RightToLeft | Parsing.RegexCompiledOption);
+            LuaNumberRegex = new Regex(@"\b\d+[\.]?\d*([eE]\-?\d+)?[lLdDfF]?\b|\b0x[a-fA-F\d]+\b",
+                                           Parsing.RegexCompiledOption);
+            LuaKeywordRegex =
+                new Regex(
+                    @"\b(and|break|do|else|elseif|end|false|for|function|if|in|local|nil|not|or|repeat|return|then|true|until|while)\b",
+                    Parsing.RegexCompiledOption);
+
+            LuaFunctionsRegex =
+                new Regex(
+                    @"\b(assert|collectgarbage|dofile|error|getfenv|getmetatable|ipairs|load|loadfile|loadstring|module|next|pairs|pcall|print|rawequal|rawget|rawset|require|select|setfenv|setmetatable|tonumber|tostring|type|unpack|xpcall)\b",
+                    Parsing.RegexCompiledOption);
+        }
+
+		public void AutoIndentNeeded(FastColoredTextBox fctb, AutoIndentEventArgs args) {
+            //end of block
+            if (Regex.IsMatch(args.LineText, @"^\s*(end|until)\b"))
+            {
+                args.Shift = -args.TabLength;
+                args.ShiftNextLines = -args.TabLength;
+                return;
+            }
+            // then ...
+            if (Regex.IsMatch(args.LineText, @"\b(then)\s*\S+"))
+                return;
+            //start of operator block
+            if (Regex.IsMatch(args.LineText, @"^\s*(function|do|for|while|repeat|if)\b"))
+            {
+                args.ShiftNextLines = args.TabLength;
+                return;
+            }
+
+            //Statements else, elseif, case etc
+            if (Regex.IsMatch(args.LineText, @"^\s*(else|elseif)\b", RegexOptions.IgnoreCase))
+            {
+                args.Shift = -args.TabLength;
+                return;
+            }
+		}
+
+		public void SyntaxHighlight(Range range) {
+            range.tb.CommentPrefix = "--";
+            range.tb.LeftBracket = '(';
+            range.tb.RightBracket = ')';
+            range.tb.LeftBracket2 = '{';
+            range.tb.RightBracket2 = '}';
+            range.tb.BracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+            //clear style of changed range
+            range.ClearStyle(StringStyle, CommentStyle, NumberStyle, KeywordStyle, FunctionsStyle);
+            //
+            if (LuaStringRegex == null)
+                InitLuaRegex();
+            //string highlighting
+            range.SetStyle(StringStyle, LuaStringRegex);
+            //comment highlighting
+            range.SetStyle(CommentStyle, LuaCommentRegex1);
+            range.SetStyle(CommentStyle, LuaCommentRegex2);
+            range.SetStyle(CommentStyle, LuaCommentRegex3);
+            //number highlighting
+            range.SetStyle(NumberStyle, LuaNumberRegex);
+            //keyword highlighting
+            range.SetStyle(KeywordStyle, LuaKeywordRegex);
+            //functions highlighting
+            range.SetStyle(FunctionsStyle, LuaFunctionsRegex);
+            //clear folding markers
+            range.ClearFoldingMarkers();
+            //set folding markers
+            range.SetFoldingMarkers("{", "}"); //allow to collapse brackets block
+            range.SetFoldingMarkers(@"--\[\[", @"\]\]"); //allow to collapse comment block
+		}
+
+	}
+}
