+++ FastColoredTextBox/Language/Php.cs
@@ -1,0 +1,120 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Text.RegularExpressions;
+using FastColoredTextBoxNS.Language2;
+
+namespace FastColoredTextBoxNS.Language2
+{
+	public class Php : ILanguage
+	{
+		private readonly Style StringStyle = LanguageStyles.Red;
+		private readonly Style CommentStyle = LanguageStyles.Green;
+		private readonly Style NumberStyle = LanguageStyles.Red;
+		private readonly Style VariableStyle = LanguageStyles.Maroon;
+		private readonly Style KeywordStyle = LanguageStyles.Green;
+		private readonly Style KeywordStyle2 = LanguageStyles.Blue;
+		private readonly Style KeywordStyle3 = LanguageStyles.Gray;
+
+        protected Regex PHPCommentRegex1,
+                      PHPCommentRegex2,
+                      PHPCommentRegex3;
+
+        protected Regex PHPKeywordRegex1,
+                      PHPKeywordRegex2,
+                      PHPKeywordRegex3;
+
+        protected Regex PHPNumberRegex;
+        protected Regex PHPStringRegex;
+        protected Regex PHPVarRegex;
+
+
+		public Php() {
+			InitPHPRegex();
+		}
+        protected void InitPHPRegex()
+        {
+            PHPStringRegex = new Regex(@"""""|''|"".*?[^\\]""|'.*?[^\\]'", Parsing.RegexCompiledOption);
+            PHPNumberRegex = new Regex(@"\b\d+[\.]?\d*\b", Parsing.RegexCompiledOption);
+            PHPCommentRegex1 = new Regex(@"(//|#).*$", RegexOptions.Multiline | Parsing.RegexCompiledOption);
+            PHPCommentRegex2 = new Regex(@"(/\*.*?\*/)|(/\*.*)", RegexOptions.Singleline | Parsing.RegexCompiledOption);
+            PHPCommentRegex3 = new Regex(@"(/\*.*?\*/)|(.*\*/)",
+                                         RegexOptions.Singleline | RegexOptions.RightToLeft | Parsing.RegexCompiledOption);
+            PHPVarRegex = new Regex(@"\$[a-zA-Z_\d]*\b", Parsing.RegexCompiledOption);
+            PHPKeywordRegex1 =
+                new Regex(
+                    @"\b(die|echo|empty|exit|eval|include|include_once|isset|list|require|require_once|return|print|unset)\b",
+                    Parsing.RegexCompiledOption);
+            PHPKeywordRegex2 =
+                new Regex(
+                    @"\b(abstract|and|array|as|break|case|catch|cfunction|class|clone|const|continue|declare|default|do|else|elseif|enddeclare|endfor|endforeach|endif|endswitch|endwhile|extends|final|for|foreach|function|global|goto|if|implements|instanceof|interface|namespace|new|or|private|protected|public|static|switch|throw|try|use|var|while|xor)\b",
+                    Parsing.RegexCompiledOption);
+            PHPKeywordRegex3 = new Regex(@"__CLASS__|__DIR__|__FILE__|__LINE__|__FUNCTION__|__METHOD__|__NAMESPACE__",
+                                         Parsing.RegexCompiledOption);
+        }
+		public void AutoIndentNeeded(FastColoredTextBox fctb, AutoIndentEventArgs args) {
+            /*
+            FastColoredTextBox tb = sender as FastColoredTextBox;
+            tb.CalcAutoIndentShiftByCodeFolding(sender, args);*/
+            //block {}
+            if (Regex.IsMatch(args.LineText, @"^[^""']*\{.*\}[^""']*$"))
+                return;
+            //start of block {}
+            if (Regex.IsMatch(args.LineText, @"^[^""']*\{"))
+            {
+                args.ShiftNextLines = args.TabLength;
+                return;
+            }
+            //end of block {}
+            if (Regex.IsMatch(args.LineText, @"}[^""']*$"))
+            {
+                args.Shift = -args.TabLength;
+                args.ShiftNextLines = -args.TabLength;
+                return;
+            }
+            //is unclosed operator in previous line ?
+            if (Regex.IsMatch(args.PrevLineText, @"^\s*(if|for|foreach|while|[\}\s]*else)\b[^{]*$"))
+                if (!Regex.IsMatch(args.PrevLineText, @"(;\s*$)|(;\s*//)")) //operator is unclosed
+                {
+                    args.Shift = args.TabLength;
+                    return;
+                }
+		}
+
+		public void SyntaxHighlight(Range range) {
+            range.tb.CommentPrefix = "//";
+            range.tb.LeftBracket = '(';
+            range.tb.RightBracket = ')';
+            range.tb.LeftBracket2 = '{';
+            range.tb.RightBracket2 = '}';
+            range.tb.BracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+            //clear style of changed range
+            range.ClearStyle(StringStyle, CommentStyle, NumberStyle, VariableStyle, KeywordStyle, KeywordStyle2,
+                             KeywordStyle3);
+            //
+            if (PHPStringRegex == null)
+                InitPHPRegex();
+            //string highlighting
+            range.SetStyle(StringStyle, PHPStringRegex);
+            //comment highlighting
+            range.SetStyle(CommentStyle, PHPCommentRegex1);
+            range.SetStyle(CommentStyle, PHPCommentRegex2);
+            range.SetStyle(CommentStyle, PHPCommentRegex3);
+            //number highlighting
+            range.SetStyle(NumberStyle, PHPNumberRegex);
+            //var highlighting
+            range.SetStyle(VariableStyle, PHPVarRegex);
+            //keyword highlighting
+            range.SetStyle(KeywordStyle, PHPKeywordRegex1);
+            range.SetStyle(KeywordStyle2, PHPKeywordRegex2);
+            range.SetStyle(KeywordStyle3, PHPKeywordRegex3);
+
+            //clear folding markers
+            range.ClearFoldingMarkers();
+            //set folding markers
+            range.SetFoldingMarkers("{", "}"); //allow to collapse brackets block
+            range.SetFoldingMarkers(@"/\*", @"\*/"); //allow to collapse comment block
+		}
+
+	}
+}
