+++ FastColoredTextBox/Range.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Drawing;
 using System.Text.RegularExpressions;
 using System.Collections.Generic;
+using FastColoredTextBoxNS.Language2;
 
 namespace FastColoredTextBoxNS
 {
@@ -701,7 +702,7 @@ namespace FastColoredTextBoxNS
         public void SetStyle(StyleIndex styleLayer, string regexPattern, RegexOptions options)
         {
             if (Math.Abs(Start.iLine - End.iLine) > 1000)
-                options |= SyntaxHighlighter.RegexCompiledOption;
+                options |= Parsing.RegexCompiledOption;
             //
             foreach (var range in GetRanges(regexPattern, options))
                 range.SetStyle(styleLayer);
@@ -752,7 +753,7 @@ namespace FastColoredTextBoxNS
         /// <param name="finishFoldingPattern">Pattern for finish folding line</param>
         public void SetFoldingMarkers(string startFoldingPattern, string finishFoldingPattern)
         {
-            SetFoldingMarkers(startFoldingPattern, finishFoldingPattern, SyntaxHighlighter.RegexCompiledOption);
+            SetFoldingMarkers(startFoldingPattern, finishFoldingPattern, Parsing.RegexCompiledOption);
         }
 
         /// <summary>
